Fixed warning (Rolandas)
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@4744 3231ed4b-2112-0410-84e0-8c3944a40730

+++ skin-installer/window.c
@@ -101,7 +101,7 @@ INT_PTR CALLBACK ExtractProc(HWND hwndDlg, UINT msg, WPARAM wParam, LPARAM lPara
 				SetBkMode(lpDrawItemStruct->hDC, TRANSPARENT);
 				SetTextColor(lpDrawItemStruct->hDC, RGB(0, 0, 0));
 				GetDlgItemText(hwndDlg, (UINT)wParam, buf, 256);
-				DrawText(lpDrawItemStruct->hDC, buf, wcslen(buf), &lpDrawItemStruct->rcItem, DT_LEFT);
+				DrawText(lpDrawItemStruct->hDC, buf, (int)wcslen(buf), &lpDrawItemStruct->rcItem, DT_LEFT);
 			}
 			break;
 		case WM_COMMAND:
